Reorganised consumer & utils structure
