+++ mindtrace/jobs/mindtrace/jobs/__init__.py
@@ -1,20 +1,30 @@
-from mindtrace.jobs.types import BackendType, ExecutionStatus, Job, JobSchema, JobInput, JobOutput
-from mindtrace.jobs.consumer import Consumer
+from mindtrace.jobs.types.jobs import Job, JobInput, JobOutput, JobSchema, BackendType, ExecutionStatus
+from mindtrace.jobs.consumers.consumer import Consumer
+from mindtrace.jobs.local.consumer_backend import LocalConsumerBackend
+from mindtrace.jobs.redis.consumer_backend import RedisConsumerBackend
+from mindtrace.jobs.rabbitmq.consumer_backend import RabbitMQConsumerBackend
 from mindtrace.jobs.orchestrator import Orchestrator
 from mindtrace.jobs.local.client import LocalClient
 from mindtrace.jobs.redis.client import RedisClient
 from mindtrace.jobs.rabbitmq.client import RabbitMQClient
+from mindtrace.jobs.utils.checks import ifnone
 
 __all__ = [
     'BackendType',
     'Consumer',
     'ExecutionStatus', 
     'Job', 
-    'JobSchema', 
     'JobInput', 
     'JobOutput', 
     'LocalClient', 
     'Orchestrator',
     'RabbitMQClient',
-    'RedisClient', 
+    'RedisClient',
+    'JobSchema',
+    'BackendType',
+    'ExecutionStatus',
+    'ifnone',
+    'LocalConsumerBackend',
+    'RedisConsumerBackend',
+    'RabbitMQConsumerBackend',
 ]
