+++ mindtrace/jobs/mindtrace/jobs/rabbitmq/consumer_backend.py
@@ -2,7 +2,7 @@ import traceback
 from typing import Optional, Callable
 
 from mindtrace.jobs.base.consumer_base import ConsumerBackendBase
-from mindtrace.jobs.utils import ifnone
+from mindtrace.jobs.utils.checks import ifnone
 
 class RabbitMQConsumerBackend(ConsumerBackendBase):
     """RabbitMQ consumer backend with improved consumption logic."""
