--- mindtrace/jobs/mindtrace/jobs/utils.py
@@ -1,6 +1,0 @@
-from typing import Any
-
-
-def ifnone(val: Any, default: Any) -> Any:
-    """Return default if val is None, otherwise return val."""
-    return default if val is None else val
